pyproject.toml
[tool.poetry]
name = "dema-ai-code-challenge"
version = "0.1.0"
description = ""
authors = ["Roberto Paredes <rdparedessalazar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
pandas = "^2.1.3"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"